Minor changes in zigzag and rle lib. Modified data types

+++ labwork3/src/lib/dct/rle/rle.cpp
@@ -6,15 +6,15 @@
 #include <tuple>
 #include "rle.h"
 
-rle::rle(vector<int> vect){
+rle::rle(vector<short> vect){
     rle::vect = vect;
 }
 
-rle::rle(vector<tuple<int, int>> vect){
+rle::rle(vector<tuple<short, short>> vect){
     rle::vect_rle = vect;
 }
 
-vector<int> rle::rm_fin_zeros(){
+vector<short> rle::rm_fin_zeros(){
 
     int idx=-1,count = 0;
 
@@ -31,21 +31,20 @@ vector<int> rle::rm_fin_zeros(){
     else
         rle::num_zeros = vect.size() - idx;
 
-    vector<int> res(vect.begin(),vect.begin()+idx);
+    vector<short> res(vect.begin(),vect.begin()+idx);
     return res;
 
 }
 
-vector<int> rle::reverse_rle_zeros(int num_zeros){
+vector<short> rle::reverse_rle_zeros(int num_zeros){
 
-    vector<int> vect_null(num_zeros, 0);
+    vector<short> vect_null(num_zeros, 0);
     vect.insert( vect.end(), vect_null.begin(), vect_null.end());
     return vect;
 
 }
 
-
-vector<int> rle::it_zero_rm(vector<int> vec) {
+vector<short> rle::it_zero_rm(vector<short> vec) {
     int idx = -1, count = 0;
 
     for(int x : vec){
@@ -53,17 +52,17 @@ vector<int> rle::it_zero_rm(vector<int> vec) {
         else if(x!=0) idx=-1;
         count++;
     }
-    vector<int> res(vec.begin(), vec.begin()+idx);
+    vector<short> res(vec.begin(), vec.begin()+idx);
     return res;
 }
 
 
-vector<int> rle::rec_zero_rm(vector<int> tmp_vec){
+vector<short> rle::rec_zero_rm(vector<short> tmp_vec){
     int middle_pnt = int(tmp_vec.size()/2);
 
     if(tmp_vec.at(middle_pnt)==0){
-        vector<int> tmp_post(tmp_vec.begin()+middle_pnt, tmp_vec.end());
-        vector<int> tmp_pre(tmp_vec.begin(),tmp_vec.begin()+middle_pnt);
+        vector<short> tmp_post(tmp_vec.begin()+middle_pnt, tmp_vec.end());
+        vector<short> tmp_pre(tmp_vec.begin(),tmp_vec.begin()+middle_pnt);
 
         if(accumulate(tmp_post.begin(), tmp_post.end(), 0) == 0)
             return rle::rec_zero_rm(tmp_pre);
@@ -74,8 +73,8 @@ vector<int> rle::rec_zero_rm(vector<int> tmp_vec){
     return rle::it_zero_rm(tmp_vec);
 }
 
-vector<tuple<int, int>> rle::get_rle(){
-    vector<tuple<int, int>> tmp(vect.size()+1,std::make_tuple(0,0));
+vector<tuple<short, short>> rle::get_rle(){
+    vector<tuple<short, short>> tmp(vect.size()+1,std::make_tuple(0,0));
     int current_val = vect[0];
     int count_current_val = 1;
     int idx = 0;
@@ -98,8 +97,8 @@ vector<tuple<int, int>> rle::get_rle(){
     return tmp;
 }
 
-vector<tuple<int, int>> rle::get_variant_rle(){
-    vector<tuple<int, int>> tmp(vect.size()+1,std::make_tuple(0,0));
+vector<tuple<short, short>> rle::get_variant_rle(){
+    vector<tuple<short, short>> tmp(vect.size()+1,std::make_tuple(0,0));
     int count_zeros = 0;
     int idx = 0;
     auto it = vect.begin();
@@ -119,33 +118,33 @@ vector<tuple<int, int>> rle::get_variant_rle(){
 
 int rle::get_vector_size() {
     int count=0;
-    for(tuple<int,int>t:vect_rle)
+    for(tuple<short,short>t:vect_rle)
         count+=std::get<1>(t);
     return count;
 }
 
 int rle::get_variant_vector_size() {
     int count=0;
-    for(tuple<int,int>t:vect_rle)
+    for(tuple<short,short>t:vect_rle)
         count+=std::get<0>(t) + 1;
     return count -1;
 }
 
-vector<int> rle::load_rle(){
-    vector<int> res(get_vector_size(),0);
+vector<short> rle::load_rle(){
+    vector<short> res(get_vector_size(),0);
     int idx=0;
-    for(tuple<int,int>t:vect_rle){
+    for(tuple<short,short>t:vect_rle){
         for(int i=0; i< std::get<1>(t); i++)
             res[idx++] = std::get<0>(t);
     }
     return res;
 }
 
-vector<int> rle::load_variant_rle(int n){
+vector<short> rle::load_variant_rle(int n){
     assert(n>=get_variant_vector_size());
-    vector<int> res(n,0);
+    vector<short> res(n,0);
     int idx=0;
-    for(tuple<int,int>t:vect_rle){
+    for(tuple<short,short>t:vect_rle){
         for(int i=0; i< std::get<0>(t); i++) res[idx++] = 0;
         res[idx++] = std::get<1>(t);
     }
@@ -157,8 +156,8 @@ int rle::get_num_zeros(){
     return rle::num_zeros;
 }
 
-vector<int> rle::get_vector(){return vect;}
+vector<short> rle::get_vector(){return vect;}
 
-vector<tuple<int, int>> rle::get_vector_rle(){return vect_rle;}
+vector<tuple<short, short>> rle::get_vector_rle(){return vect_rle;}
